requirements.txt
click
anyscale
slackclient
boto3
PyGithub
